mcp/requirements.txt
fastmcp
uvicorn
python-dotenv
pyopensky
requests